infra: finalizing some steps before releasing
