+++ .github/workflows/wheels.yml
@@ -52,10 +52,38 @@ jobs:
           pip install twine
           twine upload dist/* --repository pypi --verbose
 
+  build-arm-wheels:
+    name: Build ARM-based wheels for ${{ matrix.python }}-${{ matrix.platform }}
+    runs-on: ubuntu-latest
+    if: "!github.event.repository.fork" # Don't run on fork repository
+    timeout-minutes: 90
+    strategy:
+      fail-fast: false
+      matrix:
+        python: ["cp38", "cp39", "cp310", "cp311"]
+        platform: [manylinux_aarch64, macosx_arm64]
+    env:
+      CIBW_BUILD: ${{ matrix.python }}-${{ matrix.platform[1] }}
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 0 # fetch all tags and branches
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v2
+        with:
+          platforms: all
+      - name: Build wheels
+        uses: pypa/cibuildwheel@v2.12.0
+      - uses: actions/upload-artifact@v3
+        with:
+          name: ${{ matrix.python }}-${{ startsWith(matrix.platform, 'macosx') && 'macosx' || matrix.platform }}
+          path: ./wheelhouse/*.whl
+
   build-wheel:
     name: Build wheels for python${{ matrix.python-version }} (${{ matrix.os }})
     runs-on: ${{ matrix.os }}
-    timeout-minutes: 20
+    timeout-minutes: 90
     if: github.repository_owner == 'aarnphm' # Don't run on fork repository
     strategy:
       fail-fast: false
