+++ Makefile
@@ -29,13 +29,13 @@ endif
 context.o: src/whispercpp/context.cc src/whispercpp/context.h
 	$(CXX) $(CXXFLAGS) -o src/whispercpp/context.o -c src/whispercpp/context.cc
 
-api_export.o: src/whispercpp/api_export.cc
-	$(CXX) $(CXXFLAGS) -o src/whispercpp/api_export.o -c src/whispercpp/api_export.cc
+api_cpp2py_export.o: src/whispercpp/api_cpp2py_export.cc
+	$(CXX) $(CXXFLAGS) -o src/whispercpp/api_cpp2py_export.o -c src/whispercpp/api_cpp2py_export.cc
 
-api: api_export.o context.o
+api: api_cpp2py_export.o context.o
 	@echo "Building pybind11 extension..."
 	@cd ./extern/whispercpp && $(MAKE) ggml.o whisper.o && mv whisper.o ggml.o ../../src/whispercpp/
-	$(CXX) $(CXXFLAGS) $(EXTRA_CXXFLAGS) -shared -o src/whispercpp/api.so src/whispercpp/*.o
+	$(CXX) $(CXXFLAGS) $(EXTRA_CXXFLAGS) -shared -o src/whispercpp/api_cpp2py_export.so src/whispercpp/*.o
 
 clean:
 	rm -rf **/*.o **/*.so
