+++ pyproject.toml
@@ -129,7 +129,7 @@ skip_glob = [
 
 [tool.pyright]
 pythonVersion = "3.11"
-exclude = ["bazel-*", "extern", "venv"]
+exclude = ["bazel-*", "extern", "venv", "node_modules", "nix", "setup.py", "examples/bentoml/locustfile.py"]
 typeCheckingMode = "strict"
 analysis.useLibraryCodeForTypes = true
 enableTypeIgnoreComments = true
