+++ src/whispercpp/BUILD
@@ -7,5 +7,8 @@ exports_files(glob(["*.cc"]) + glob(["*.h"]) + glob(["*.py"]) + glob(["*.pyi"]))
 py_library(
     name = "whispercpp_lib",
     srcs = glob(["*.py"]),
-    data = ["//:api.so"] + glob(["*.pyi"]),
+    data = [
+        "//:api_cpp2py_export.so",
+        "//:audio_cpp2py_export.so",
+    ] + glob(["*.pyi"]),
 )
