+++ tests/export_test.py
@@ -6,17 +6,15 @@ from pathlib import Path
 
 import pytest as p
 
-from whispercpp import api
-from whispercpp import Whisper
-from whispercpp.utils import LazyLoader
+import whispercpp as w
 
 if t.TYPE_CHECKING:
     import numpy as np
     import ffmpeg
     from numpy.typing import NDArray
 else:
-    np = LazyLoader("np", globals(), "numpy")
-    ffmpeg = LazyLoader("ffmpeg", globals(), "ffmpeg")
+    np = w.utils.LazyLoader("np", globals(), "numpy")
+    ffmpeg = w.utils.LazyLoader("ffmpeg", globals(), "ffmpeg")
 
 ROOT = Path(__file__).parent.parent
 
@@ -38,12 +36,12 @@ def preprocess(file: Path, sample_rate: int = 16000) -> NDArray[np.float32]:
 
 def test_invalid_models():
     with p.raises(RuntimeError):
-        Whisper.from_pretrained("whisper_v0.1")
+        w.Whisper.from_pretrained("whisper_v0.1")
 
 
 def test_forbid_init():
     with p.raises(RuntimeError):
-        Whisper()
+        w.Whisper()
 
 
 _EXPECTED = " And so my fellow Americans ask not what your country can do for you ask what you can do for your country"
@@ -51,7 +49,7 @@ _EXPECTED = " And so my fellow Americans ask not what your country can do for yo
 
 @p.mark.skipif(not s.which("ffmpeg"), reason="ffmpeg not found, skipping this test.")
 def test_from_pretrained():
-    m = Whisper.from_pretrained("tiny.en")
+    m = w.Whisper.from_pretrained("tiny.en")
     assert _EXPECTED == m.transcribe(preprocess(ROOT / "samples" / "jfk.wav"))
 
 
@@ -59,14 +57,14 @@ def test_from_pretrained():
 def test_load_wav_file():
     np.testing.assert_almost_equal(
         preprocess(ROOT / "samples" / "jfk.wav"),
-        api.load_wav_file(
+        w.api.load_wav_file(
             ROOT.joinpath("samples", "jfk.wav").resolve().__fspath__()
         ).mono,
     )
 
 
 def test_transcribe_from_wav():
-    m = Whisper.from_pretrained("tiny.en")
+    m = w.Whisper.from_pretrained("tiny.en")
     assert (
         m.transcribe_from_file(
             ROOT.joinpath("samples", "jfk.wav").resolve().__fspath__()
@@ -76,14 +74,14 @@ def test_transcribe_from_wav():
 
 
 def test_callback():
-    def handleNewSegment(context: api.Context, n_new: int, text: list[str]):
+    def handleNewSegment(context: w.api.Context, n_new: int, text: list[str]):
         segment = context.full_n_segments() - n_new
         while segment < context.full_n_segments():
             text.append(context.full_get_segment_text(segment))
             print(text)
             segment += 1
 
-    m = Whisper.from_pretrained("tiny.en")
+    m = w.Whisper.from_pretrained("tiny.en")
 
     text = []
     m.params.on_new_segment(handleNewSegment, text)
